dlfcn.h: Decorate protoypes with dllimport if not building the library
This enables getting imported addresses directly from the image allocation
table, which works similar to -fpie on UNIX like os.

In order for clients to be able to distinguish between the two
installation options of the dl library (dynamic and static), they must set
the preprocessor macro -DDLFCN_WIN32_SHARED if the dynamic variant is
desired, which was renamed from SHARED to avoid name clashing with other
packages using the same unspecific name.

With cmake clients this is done automatically by the configuration files
installed for find_package support.