Add function dladdr()
- fill module path, base address

- fill symbol address to have really used address

- fill symbol address for symbols located in a shared library
  by determining the name from the IAT if the related symbol
  is not imported by ordinal.

If the original address points to a thunk table entry,
which is the case if the requested symbol is not decorated
with __attribute__((dllimport)), you get the real address from
the Image Allocation Table (IAT). These retrieved addresses
are checked against the IAT address range to ensure validity.
On x86_64 the offset after the 'jmpq' statement does
not match the size of a pointer used by 32 bits, which
requires a different approach to determine the address
from the IAT.